UI Updates: MVVM implemented

+++ HamsterDatabase/Procedures/InStockNewItem.sql
@@ -2,8 +2,8 @@
 	@productId int,
 	@productionDate datetime2,
 	@stockedDate datetime2,
-	@producedQuantity decimal,
-	@unitPrice money,
+	@producedQuantity int,
+	@unitPrice decimal,
 	@workerId int
 AS
 	INSERT INTO Stock (ProductId, ProductionDate, StockedDate, ProducedQuantity, UnitPrice, WorkerId)
